merge

+++ WADWolfram/src/main/java/com/mycompany/mavenproject1/ProjectController.java
@@ -41,7 +41,7 @@ projects.delete(p);
 
 	@RequestMapping ("/project")
 	public String loadProject(Model m){
-		int id=0;
+		long id=0;
 		String title = "title";
 		String shortDescription="shortDescription";
 		String description = "description";
@@ -56,7 +56,7 @@ projects.delete(p);
 	}
 */	
 	@RequestMapping("/pay")
-	public String donate(Model m, int projectId){
+	public String donate(Model m, long projectId){
 		//projectId es el id para reconocer al proyecto que se dona
 		
 		return "pay";
